re-debug the itemDetails. Refs UIREQ-55
itemRecord and item, plain,
Were both perhaps present in selectedItem.
And itemBarcode contained in selectedItem.
In in selectedItem, item, plain, is expunged, and other eternally present,
Request-New and Request-Edit are all clickable.
Passing tests will be no more an abstraction,
becoming a perpetual possibility
in a world of speculative compilation.
What might have passed and what has passed
point to ui-inventory and ui-requests; it is always my fault.
Bug reports echo in the ticket tracker
To the continuous integration server we do not like
Towards the green line we never see
Into the slack-channel. My fix echoes
thus, in your code.