read query string for user and item barcodes
Pass the query resource down to `<RequestForm>` so it can access the
query string to lookup the user and item records based on the presence
of `userBarcode` and `itemBarcode`.

Additionally, avoid `setTimeout`. This was in there previously because
redux-form's `props.change` method is async and we needed to let that
propagate first. Thing is, we _have_ the user barcode already, so we
don't actually need to wait for that propagation; we can just pass the
value directly. And now we do.

Fixes UIREQ-146