make unit tests responsible for setting state
It seems to be the case that application served up by mirage comes in
some sort of virgin state regardless of what URL we pass to
`this.visit()`. Deployed, a URL like `/requests?filter=foo` would cause
the "foo" checkbox to be ticked. Not so when running tests and passing a
fancy URL like

    /requests?filters=requestType.Holds%2CrequestType.Pages%2CrequestType.Recalls&sort=Request%20Date

appears to return the same page as passing

    /requests

What this means is that we need to use interators to affect the state
that normally we would expect to inherit through the URL.

Fixes UIREQ-210