UIREQ-304:  Missing Instance title from the Moving Request popup
The label in question had a static placeholder in the translation file for the
item title.  I replaced this with a variable, and passed in the item title from the
modal render function.  Once this change propagates out to all the translation
files, they should start showing the item title in the modal.

Refs: https://issues.folio.org/browse/UIREQ-304