UIREQ-870 allow request-move in non-English locales
Instead of translating `item.status.name` in the underlying data
structure and then displaying that "raw" value, leave the data alone and
include the translation function in the column formatter.

This prevents an NPE in `ChooseRequestTypeDialog` which uses the item's
status' name as a lookup key to determine which types of requests are
valid for the given status.

Refs UIREQ-870